# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/job": {
            "get": {
                "description": "分页获取任务列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "获取任务列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始位置",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dao.ListJobsResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "创建任务",
                "parameters": [
                    {
                        "description": "创建任务请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dao.CreateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dao.CreateJobResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/job/{job_id}": {
            "get": {
                "description": "根据job_id获取任务详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "获取任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务job_id",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dao.JobSpec"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "任务不存在",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "根据job_id更新任务信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "更新任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务job_id",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新任务请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dao.UpdateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功"
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "任务不存在",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据job_id删除任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "删除任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务job_id",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "任务不存在",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/job/{job_id}/start": {
            "put": {
                "description": "根据job_id启动任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "启动任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务job_id",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "启动成功"
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "任务不存在",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/job/{job_id}/stop": {
            "put": {
                "description": "根据job_id停止任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "停止任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务job_id",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "停止成功"
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "任务不存在",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dao.CreateJobRequest": {
            "type": "object",
            "required": [
                "input",
                "kind"
            ],
            "properties": {
                "detect": {
                    "$ref": "#/definitions/dao.DetectOptions"
                },
                "deviceId": {
                    "type": "integer"
                },
                "input": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/model.JobKind"
                },
                "query": {
                    "type": "string"
                },
                "videoSegment": {
                    "$ref": "#/definitions/dao.VideoSegmentOptions"
                },
                "workflowId": {
                    "type": "integer"
                }
            }
        },
        "dao.CreateJobResponse": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "dao.DetectOptions": {
            "type": "object",
            "required": [
                "modelName"
            ],
            "properties": {
                "confThreshold": {
                    "type": "number"
                },
                "interval": {
                    "type": "integer"
                },
                "iouThreshold": {
                    "type": "number"
                },
                "labels": {
                    "type": "string"
                },
                "modelName": {
                    "type": "string"
                }
            }
        },
...
        "dao.DeviceSpec": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lastPingTime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "registerTime": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dao.JobSpec": {
            "type": "object",
            "required": [
                "createTime",
                "input",
                "kind",
                "status",
                "updateTime",
                "uuid"
            ],
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "detect": {
                    "$ref": "#/definitions/dao.DetectOptions"
                },
                "device": {
                    "$ref": "#/definitions/dao.DeviceSpec"
                },
                "id": {
                    "type": "integer"
                },
                "input": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/model.JobKind"
                },
                "query": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.JobStatus"
                },
                "updateTime": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "videoSegment": {
                    "$ref": "#/definitions/dao.VideoSegmentOptions"
                },
                "workflowId": {
                    "type": "integer"
                }
            }
        },
...
        "dao.ListJobsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.JobSpec"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dao.UpdateJobRequest": {
            "type": "object",
            "properties": {
                "detect": {
                    "$ref": "#/definitions/dao.DetectOptions"
                },
                "deviceId": {
                    "type": "integer"
                },
                "input": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "videoSegment": {
                    "$ref": "#/definitions/dao.VideoSegmentOptions"
                },
                "workflowId": {
                    "type": "integer"
                }
            }
        },
...
        "dao.VideoSegmentOptions": {
            "type": "object",
            "properties": {
                "interval": {
                    "type": "integer"
                }
            }
        },
...
        "model.JobKind": {
            "type": "string",
            "enum": [
                "detect",
                "video_segment"
            ],
            "x-enum-varnames": [
                "JobKindDetect",
                "JobKindVideoSegment"
            ]
        },
        "model.JobStatus": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "JobStatusStopped",
                "JobStatusRunning"
            ]
        },
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        }
    }
}